Add config for node arrangement algorithm
- And a placeholder for the next algorithm.
